+++ src/ddzy/CHANGELOG.md
@@ -269,3 +269,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 重构项目
   - 拆分各工具方法为单个模块, 更明朗的结构
   - 重构之前的版本移至: [https://github.com/ddzy/ts-utility-plugins/tree/branch/rel-1.0](https://github.com/ddzy/ts-utility-plugins/tree/branch/rel-1.0)
+
+## ✍ [2019-8-23]
+
+### Added
+
+- 新增`utility/date`区块
+- utility/date
+  - 新增`now`方法
+- utility/array
+  - 新增`castArray`方法